// src/app/models/pagofacil.model.ts
export class Pagofacilrequest {
    nombre: String;
    apellidos: String;
    numeroTarjeta: String;
    cvt: String;
    cp: String;
    mesExpiracion: String;
    anyoExpiracion: String;
    monto: String;
    idSucursal: String;
    idUsuario: String;
    idServicio: String;
    email: String;
    telefono: String;
    celular: String;
    calleyNumero: String;
    colonia: String;
    municipio: String;
    estado: String;
    pais: String;
    param1: String;
    param2: String;
    param3: String;
    param4: String;
    param5: String
}